+++ test/libsolidity/syntaxTests/functionTypes/function_type_constructor.sol
@@ -1,0 +1,7 @@
+contract C {
+    // Fool parser into parsing a constructor as a function type.
+    constructor() x;
+}
+// ----
+// Warning: (83-99): Modifiers of functions without implementation are ignored.
+// DeclarationError: (97-98): Undeclared identifier.