+++ test/libsolidity/syntaxTests/functionTypes/internal_function_returned_from_public_function.sol
@@ -1,0 +1,7 @@
+// It should not be possible to return internal functions from external functions.
+contract C {
+    function f() public returns (function(uint) internal returns (uint) x) {
+    }
+}
+// ----
+// TypeError: (129-169): Internal or recursive type is not allowed for public or external functions.