+++ test/libsolidity/syntaxTests/modifiers/legal_modifier_override.sol
@@ -1,0 +1,2 @@
+contract A { modifier mod(uint a) { _; } }
+contract B is A { modifier mod(uint a) { _; } }